# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CipherPay Fintech API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List user transactions with filters
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get single transaction details
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── BANK ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Transaction:
      type: object
      required: [id, userId, type, amount, status, reference, description, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        amount:
          type: number
        fee:
          type: ["number", "null"]
        status:
          type: string
        reference:
          type: string
        description:
          type: string
        metadata:
          type: ["string", "null"]
        balanceBefore:
          type: ["number", "null"]
        balanceAfter:
          type: ["number", "null"]
        createdAt:
          type: string
    TransactionListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── BANK ────────────────────────────────────────────────────────────────
...